Linear Regression

Estimates a vector for the model coefficients. It also computes model evaluation metrics.

Usage,

our_lm(y, X, alpha = 0.05)

Arguments

y

A vector of length n that represents the response variable of interest that we would like to explain/predict.

X

A matrix of dimension n x p that represents the independent (explanatory) variables.

alpha

A numeric (double) used to denote the significance level.

Value

A list containing the following attributes:

beta

Estimated coefficient vector

sigma2

Estimated variance

variance_beta

Estimated variance for each beta

r_squared

Coefficient of determination

mallow

Mallow's C_p

f_stat

F-statistic

p_value

Determined by F-statistic

ci

Interval of values based on alpha

res

Model Residuals

preds

Model Predictions

Author

Clayton Haley

Ukamaka Victoria Nnyaba

Robert Fett

Examples

x <- matrix(rnorm(400), ncol = 4) # so this is a 100x4 predictor matrix
y <- rnorm(100)
model <- our_lm(y, x, alpha = 0.05)
#> Warning: NaNs produced
#> Warning: NaNs produced
model
#> $beta
#> [1] -0.08711383 -0.06661126 -0.00379784  0.02205282
#> 
#> $sigma2
#>          [,1]
#> [1,] 5.969653
#> 
#> $variance_beta
#>              [,1]          [,2]          [,3]         [,4]
#> [1,]  0.061983935  0.0081683191 -0.0081144049 -0.008167658
#> [2,]  0.008168319  0.0659065061  0.0003979464 -0.013027538
#> [3,] -0.008114405  0.0003979464  0.0441198068 -0.001991332
#> [4,] -0.008167658 -0.0130275376 -0.0019913321  0.088321646
#> 
#> $r_squared
#> [1] -0.01503868
#> 
#> $mallow
#>          [,1]
#> [1,] 191.0289
#> 
#> $f_stat
#> [1] 0.9232096
#> 
#> $p_value
#> [1] 0.4327087
#> 
#> $ci
#>  [1] -0.57507769 -0.24375036         NaN         NaN -0.26425293 -0.56977836
#>  [7] -0.04289636         NaN         NaN -0.10570978 -0.41548270         NaN
#> [13]         NaN         NaN         NaN -0.56042805  0.40085003  0.11052784
#> [19]         NaN         NaN  0.09002527  0.43655584  0.03530068         NaN
#> [25]         NaN -0.02751273  0.40788702         NaN         NaN         NaN
#> [31]         NaN  0.60453369
#> 
#> $res
#>               [,1]
#>   [1,]  1.94453062
#>   [2,] -1.96015250
#>   [3,] -0.70709565
#>   [4,] -1.09359738
#>   [5,] -1.29090007
#>   [6,] -1.00614674
#>   [7,]  0.91977257
#>   [8,] -0.26613831
#>   [9,] -0.39335212
#>  [10,] -0.37367508
#>  [11,] -1.70861966
#>  [12,] -0.19146132
#>  [13,] -0.84409095
#>  [14,] -0.59540868
#>  [15,] -2.20729704
#>  [16,] -1.39141864
#>  [17,]  0.32808363
#>  [18,]  0.78946877
#>  [19,]  0.66461153
#>  [20,]  0.37752086
#>  [21,] -0.28221027
#>  [22,] -0.74358636
#>  [23,]  0.87826804
#>  [24,]  0.50388272
#>  [25,]  1.20175952
#>  [26,] -0.65136855
#>  [27,] -1.25543211
#>  [28,]  1.18787118
#>  [29,] -2.35101350
#>  [30,] -1.63301180
#>  [31,] -0.71117874
#>  [32,]  1.35942795
#>  [33,]  0.51160014
#>  [34,] -1.36139497
#>  [35,]  0.24362236
#>  [36,]  0.59130414
#>  [37,]  0.34913327
#>  [38,] -0.18252302
#>  [39,] -0.55767829
#>  [40,]  0.72105160
#>  [41,] -0.36451543
#>  [42,] -1.22651695
#>  [43,]  0.70259882
#>  [44,] -0.39493452
#>  [45,]  0.88572408
#>  [46,] -1.82774187
#>  [47,]  0.93098088
#>  [48,]  1.19459330
#>  [49,] -0.47732187
#>  [50,] -0.76323696
#>  [51,]  0.84118670
#>  [52,] -0.70167329
#>  [53,]  1.02924043
#>  [54,] -2.32425147
#>  [55,]  1.64019756
#>  [56,]  1.91040219
#>  [57,]  0.12261458
#>  [58,]  1.22745427
#>  [59,]  1.30415755
#>  [60,]  0.87530223
#>  [61,] -0.48700882
#>  [62,] -0.33441966
#>  [63,] -2.05725149
#>  [64,] -2.85999720
#>  [65,]  3.11500886
#>  [66,] -1.64888606
#>  [67,]  0.76607922
#>  [68,] -1.45686372
#>  [69,]  0.19810918
#>  [70,] -0.02133658
#>  [71,]  0.11542991
#>  [72,]  0.19212108
#>  [73,]  1.05770714
#>  [74,] -1.26403681
#>  [75,] -0.74669877
#>  [76,] -0.24062422
#>  [77,] -1.35800130
#>  [78,] -2.04516026
#>  [79,]  0.91178237
#>  [80,]  0.28047168
#>  [81,] -1.12018809
#>  [82,]  2.87038184
#>  [83,]  0.35249993
#>  [84,]  1.56031625
#>  [85,] -1.95942859
#>  [86,] -0.72280249
#>  [87,]  0.25009654
#>  [88,]  1.10086621
#>  [89,] -0.36828857
#>  [90,] -0.29275518
#>  [91,] -1.04596890
#>  [92,] -1.74766342
#>  [93,] -2.17109918
#>  [94,] -1.00416970
#>  [95,] -0.28094271
#>  [96,]  0.58309059
#>  [97,]  1.10567955
#>  [98,] -1.06354417
#>  [99,] -0.15954496
#> [100,]  1.00424059
#> 
#> $preds
#>                [,1]
#>   [1,]  0.186315472
#>   [2,]  0.141424152
#>   [3,]  0.069592652
#>   [4,] -0.085857162
#>   [5,] -0.032080760
#>   [6,]  0.006614471
#>   [7,] -0.111645908
#>   [8,]  0.179090252
#>   [9,] -0.050645139
#>  [10,] -0.153751855
#>  [11,] -0.065438560
#>  [12,] -0.063487720
#>  [13,] -0.081378609
#>  [14,] -0.108450966
#>  [15,] -0.210916980
#>  [16,] -0.004093397
#>  [17,]  0.065174868
#>  [18,]  0.014177301
#>  [19,]  0.082144003
#>  [20,]  0.068647842
#>  [21,] -0.163944372
#>  [22,] -0.112369791
#>  [23,]  0.018112142
#>  [24,]  0.082195433
#>  [25,] -0.231044109
#>  [26,] -0.162482928
#>  [27,]  0.020268845
#>  [28,]  0.055248488
#>  [29,] -0.013577756
#>  [30,]  0.031297515
#>  [31,] -0.133605389
#>  [32,]  0.165638075
#>  [33,] -0.047723049
#>  [34,]  0.008616576
#>  [35,] -0.068202418
#>  [36,]  0.058666180
#>  [37,]  0.017143407
#>  [38,] -0.181045718
#>  [39,]  0.138788331
#>  [40,] -0.150443189
#>  [41,]  0.036269219
#>  [42,]  0.085132004
#>  [43,] -0.069303806
#>  [44,] -0.025531751
#>  [45,]  0.123058488
#>  [46,] -0.024777953
#>  [47,] -0.062587411
#>  [48,] -0.010501991
#>  [49,]  0.072072401
#>  [50,] -0.162134261
#>  [51,]  0.077603784
#>  [52,]  0.166310371
#>  [53,] -0.078164797
#>  [54,] -0.076267643
#>  [55,]  0.006576835
#>  [56,] -0.091004337
#>  [57,]  0.154027682
#>  [58,] -0.047503895
#>  [59,]  0.002224033
#>  [60,] -0.013174695
#>  [61,]  0.054343706
#>  [62,]  0.076159069
#>  [63,] -0.009615008
#>  [64,]  0.046824693
#>  [65,] -0.213331646
#>  [66,] -0.115748683
#>  [67,]  0.117392618
#>  [68,]  0.123586927
#>  [69,]  0.016687771
#>  [70,] -0.087925586
#>  [71,]  0.146800861
#>  [72,] -0.044754234
#>  [73,]  0.008445795
#>  [74,]  0.044271913
#>  [75,] -0.145580520
#>  [76,]  0.130775203
#>  [77,] -0.011504683
#>  [78,]  0.091302324
#>  [79,]  0.124455430
#>  [80,] -0.059494057
#>  [81,]  0.155953327
#>  [82,]  0.002779784
#>  [83,] -0.167073149
#>  [84,] -0.006014427
#>  [85,] -0.013746871
#>  [86,]  0.028186592
#>  [87,]  0.211785641
#>  [88,]  0.131632362
#>  [89,]  0.115595378
#>  [90,] -0.119212398
#>  [91,] -0.005834425
#>  [92,]  0.049176237
#>  [93,]  0.009255693
#>  [94,]  0.041344146
#>  [95,] -0.210079449
#>  [96,] -0.052453069
#>  [97,]  0.028523959
#>  [98,]  0.035724151
#>  [99,]  0.091811214
#> [100,] -0.042509682
#>