Linear Regression

Estimates a vector for the model coefficients. It also computes model evaluation metrics.

Usage,

our_lm(y, X, alpha = 0.05)

Arguments

y

A vector of length n that represents the response variable of interest that we would like to explain/predict.

X

A matrix of dimension n x p that represents the independent (explanatory) variables.

alpha

A numeric (double) used to denote the significance level.

Value

A list containing the following attributes:

beta

Estimated coefficient vector

sigma2

Estimated variance

variance_beta

Estimated variance for each beta

r_squared

Coefficient of determination

mallow

Mallow's C_p

f_stat

F-statistic

p_value

Determined by F-statistic

ci

Interval of values based on alpha

res

Model Residuals

preds

Model Predictions

Author

Clayton Haley

Ukamaka Victoria Nnyaba

Robert Fett

Examples

x <- matrix(rnorm(400), ncol = 4) # so this is a 100x4 predictor matrix
y <- rnorm(100)
model <- our_lm(y, x, alpha = 0.05)
#> Warning: NaNs produced
#> Warning: NaNs produced
model
#> $beta
#> [1]  0.059878542 -0.150553401 -0.036889300  0.005331131
#> 
#> $sigma2
#>          [,1]
#> [1,] 4.485036
#> 
#> $variance_beta
#>              [,1]         [,2]         [,3]         [,4]
#> [1,]  0.044089292 -0.003599006 -0.002153518 -0.001873970
#> [2,] -0.003599006  0.043201792  0.002772669  0.006269568
#> [3,] -0.002153518  0.002772669  0.038236616  0.001593817
#> [4,] -0.001873970  0.006269568  0.001593817  0.056267065
#> 
#> $r_squared
#> [1] 0.02421494
#> 
#> $mallow
#>          [,1]
#> [1,] 143.5211
#> 
#> $f_stat
#> [1] 0.8226727
#> 
#> $p_value
#> [1] 0.4844982
#> 
#> $ci
#>  [1] -0.351663922          NaN          NaN          NaN          NaN
#>  [6] -0.557932713 -0.140093444 -0.149860000          NaN -0.253757545
#> [11] -0.420144286 -0.072915811          NaN -0.305744532 -0.115136242
#> [16] -0.459585648  0.471421006          NaN          NaN          NaN
#> [21]          NaN  0.256825911  0.066314844  0.160522261          NaN
#> [26] -0.047349257  0.346365686  0.083578073          NaN  0.004637729
#> [31]  0.041357642  0.470247910
#> 
#> $res
#>                [,1]
#>   [1,] -0.324740154
#>   [2,]  1.137489103
#>   [3,]  1.935963328
#>   [4,] -1.755361936
#>   [5,]  0.492284998
#>   [6,]  0.083560598
#>   [7,]  1.291268520
#>   [8,] -1.271243237
#>   [9,]  0.864862615
#>  [10,] -1.998279623
#>  [11,] -1.136866057
#>  [12,]  1.233933009
#>  [13,] -1.197021541
#>  [14,]  0.690980713
#>  [15,]  0.446656171
#>  [16,]  0.924149629
#>  [17,] -1.737522596
#>  [18,]  0.203936301
#>  [19,]  0.346640606
#>  [20,] -1.106464227
#>  [21,]  1.364825891
#>  [22,] -1.312876216
#>  [23,]  0.661273551
#>  [24,]  1.538337225
#>  [25,] -1.698274635
#>  [26,] -1.263384740
#>  [27,] -0.477169600
#>  [28,] -0.845077112
#>  [29,]  0.048021056
#>  [30,] -1.441217561
#>  [31,]  1.867618566
#>  [32,] -0.235631305
#>  [33,]  0.002890058
#>  [34,] -0.811635351
#>  [35,] -0.892936419
#>  [36,]  0.489595815
#>  [37,]  1.109308798
#>  [38,] -0.966562750
#>  [39,] -0.457077135
#>  [40,]  1.378598236
#>  [41,] -0.508792380
#>  [42,] -0.504529871
#>  [43,]  1.530099856
#>  [44,]  0.685610623
#>  [45,]  0.731240800
#>  [46,]  0.055694265
#>  [47,]  0.025496536
#>  [48,] -0.026560968
#>  [49,]  0.928489040
#>  [50,] -0.167530542
#>  [51,] -0.705037914
#>  [52,] -1.481838418
#>  [53,] -0.801633254
#>  [54,]  0.490377867
#>  [55,] -0.318186972
#>  [56,] -0.149577819
#>  [57,]  0.490133463
#>  [58,]  0.012744448
#>  [59,]  2.434691896
#>  [60,] -0.300351286
#>  [61,]  0.403498452
#>  [62,]  0.889222772
#>  [63,] -1.242764084
#>  [64,] -0.088604366
#>  [65,]  0.104384881
#>  [66,] -1.182462195
#>  [67,] -0.297421546
#>  [68,] -1.571584014
#>  [69,] -0.321538824
#>  [70,] -1.331098297
#>  [71,]  0.769298713
#>  [72,] -1.144904972
#>  [73,]  0.781336815
#>  [74,]  0.099291887
#>  [75,]  1.057780818
#>  [76,]  0.223183501
#>  [77,] -1.583850163
#>  [78,]  0.597876086
#>  [79,] -1.310750996
#>  [80,]  1.377747531
#>  [81,]  0.866427599
#>  [82,] -0.420427980
#>  [83,]  0.282998183
#>  [84,] -0.672685245
#>  [85,] -0.565205726
#>  [86,]  1.113182313
#>  [87,]  1.794502539
#>  [88,]  0.902228291
#>  [89,]  1.905369348
#>  [90,] -0.323386627
#>  [91,]  0.139524354
#>  [92,] -1.806102824
#>  [93,]  1.040338090
#>  [94,] -0.905073123
#>  [95,]  1.077634868
#>  [96,] -0.410767734
#>  [97,] -1.772561620
#>  [98,]  1.563042346
#>  [99,] -0.696662194
#> [100,] -0.971298003
#> 
#> $preds
#>                [,1]
#>   [1,]  0.091210435
#>   [2,]  0.070029502
#>   [3,]  0.285562950
#>   [4,]  0.115579873
#>   [5,]  0.080780589
#>   [6,] -0.161816353
#>   [7,]  0.093894226
#>   [8,]  0.121332565
#>   [9,]  0.009085173
#>  [10,]  0.123583695
#>  [11,] -0.403818293
#>  [12,] -0.067576081
#>  [13,] -0.235885950
#>  [14,]  0.159009453
#>  [15,]  0.002713719
#>  [16,]  0.006375635
#>  [17,] -0.175394721
#>  [18,]  0.006002840
#>  [19,]  0.049023984
#>  [20,] -0.181538035
#>  [21,] -0.113975717
#>  [22,]  0.048434086
#>  [23,]  0.141177791
#>  [24,]  0.147936402
#>  [25,] -0.077702461
#>  [26,] -0.050605342
#>  [27,]  0.137120955
#>  [28,] -0.016325756
#>  [29,] -0.192105365
#>  [30,]  0.267807404
#>  [31,] -0.346175006
#>  [32,] -0.046108558
#>  [33,] -0.057576917
#>  [34,] -0.133905740
#>  [35,] -0.077981745
#>  [36,] -0.286506764
#>  [37,] -0.246476155
#>  [38,]  0.092336040
#>  [39,]  0.193765587
#>  [40,]  0.054154867
#>  [41,]  0.206221165
#>  [42,] -0.155787464
#>  [43,] -0.134732064
#>  [44,]  0.081767941
#>  [45,]  0.136803068
#>  [46,]  0.174127463
#>  [47,]  0.066905048
#>  [48,] -0.142153565
#>  [49,] -0.198406410
#>  [50,]  0.042651418
#>  [51,]  0.103172933
#>  [52,]  0.207378211
#>  [53,] -0.012715847
#>  [54,] -0.148619772
#>  [55,]  0.151929789
#>  [56,] -0.219587843
#>  [57,] -0.052542688
#>  [58,]  0.030982059
#>  [59,]  0.040265407
#>  [60,] -0.061958365
#>  [61,] -0.301088309
#>  [62,] -0.046218685
#>  [63,]  0.021903460
#>  [64,]  0.133114078
#>  [65,] -0.161875019
#>  [66,] -0.090288324
#>  [67,] -0.201822729
#>  [68,]  0.075283473
#>  [69,] -0.232043656
#>  [70,]  0.014712506
#>  [71,]  0.124266012
#>  [72,]  0.228058831
#>  [73,] -0.088432933
#>  [74,]  0.047962844
#>  [75,] -0.016737416
#>  [76,] -0.106872405
#>  [77,]  0.193459806
#>  [78,]  0.180495525
#>  [79,] -0.151006085
#>  [80,] -0.325629521
#>  [81,]  0.203504837
#>  [82,]  0.024464825
#>  [83,] -0.124097190
#>  [84,]  0.250014838
#>  [85,]  0.072956441
#>  [86,]  0.173456113
#>  [87,] -0.103607205
#>  [88,]  0.474902258
#>  [89,] -0.117793853
#>  [90,] -0.177788194
#>  [91,]  0.057073162
#>  [92,]  0.171437473
#>  [93,] -0.101255949
#>  [94,] -0.147995165
#>  [95,]  0.433958743
#>  [96,]  0.032945530
#>  [97,] -0.176050832
#>  [98,]  0.149515120
#>  [99,] -0.233710642
#> [100,] -0.047194197
#>