Linear Regression

Estimates a vector for the model coefficients. It also computes model evaluation metrics.

Usage,

our_lm(y, X, alpha = 0.05)

Arguments

y

A vector of length n that represents the response variable of interest that we would like to explain/predict.

X

A matrix of dimension n x p that represents the independent (explanatory) variables.

alpha

A numeric (double) used to denote the significance level.

Value

A list containing the following attributes:

beta

Estimated coefficient vector

sigma2

Estimated variance

variance_beta

Estimated variance for each beta

r_squared

Coefficient of determination

mallow

Mallow's C_p

f_stat

F-statistic

p_value

Determined by F-statistic

ci

Interval of values based on alpha

res

Model Residuals

preds

Model Predictions

Author

Clayton Haley

Ukamaka Victoria Nnyaba

Robert Fett

Examples

x <- matrix(rnorm(400), ncol = 4) # so this is a 100x4 predictor matrix
y <- rnorm(100)
model <- our_lm(y, x, alpha = 0.05)
#> Warning: NaNs produced
#> Warning: NaNs produced
model
#> $beta
#> [1]  0.02704290 -0.05570656 -0.01398077  0.20027196
#> 
#> $sigma2
#>         [,1]
#> [1,] 4.32805
#> 
#> $variance_beta
#>              [,1]          [,2]          [,3]         [,4]
#> [1,]  0.044958858  0.0065086015 -0.0028371730 -0.001616293
#> [2,]  0.006508601  0.0495150237  0.0001983624 -0.007017270
#> [3,] -0.002837173  0.0001983624  0.0344591928  0.001844742
#> [4,] -0.001616293 -0.0070172698  0.0018447424  0.043617015
#> 
#> $r_squared
#> [1] 0.02073256
#> 
#> $mallow
#>          [,1]
#> [1,] 138.4976
#> 
#> $f_stat
#> [1] 2.373206
#> 
#> $p_value
#> [1] 0.07504186
#> 
#> $ci
#>  [1] -0.38853814 -0.21382842         NaN         NaN -0.13107896 -0.49183718
#>  [7] -0.04158513         NaN         NaN -0.08331092 -0.37781256  0.11609060
#> [13]         NaN         NaN -0.09816213 -0.20906038  0.44262395  0.10241531
#> [19]         NaN         NaN  0.18516477  0.38042407  0.01362360         NaN
#> [25]         NaN -0.02810219  0.34985102  0.28445333         NaN         NaN
#> [31]  0.07020060  0.60960430
#> 
#> $res
#>               [,1]
#>   [1,]  1.58568062
#>   [2,] -0.86977952
#>   [3,]  2.84058371
#>   [4,] -0.33805274
#>   [5,]  1.28715926
#>   [6,]  1.15854893
#>   [7,] -2.10219143
#>   [8,]  0.66465735
#>   [9,] -0.72782941
#>  [10,] -1.37423182
#>  [11,]  0.58808144
#>  [12,]  1.08931806
#>  [13,]  0.60431918
#>  [14,]  3.37648921
#>  [15,] -0.64814771
#>  [16,]  0.37061958
#>  [17,]  1.98811101
#>  [18,] -0.94072100
#>  [19,]  1.49141970
#>  [20,]  0.24918262
#>  [21,]  1.05151150
#>  [22,]  0.35564514
#>  [23,]  0.45635062
#>  [24,] -0.80292383
#>  [25,]  0.81759967
#>  [26,] -0.38931888
#>  [27,] -1.68586908
#>  [28,] -0.10028644
#>  [29,] -0.06344933
#>  [30,] -0.45664458
#>  [31,] -0.43944382
#>  [32,] -0.26117847
#>  [33,]  2.19681434
#>  [34,]  0.36079756
#>  [35,] -0.06788013
#>  [36,] -0.78337386
#>  [37,]  0.30963299
#>  [38,]  0.36964273
#>  [39,]  0.85173081
#>  [40,] -0.73848974
#>  [41,] -0.55915625
#>  [42,]  0.01874292
#>  [43,] -0.16066020
#>  [44,]  0.37140695
#>  [45,]  0.75216930
#>  [46,]  1.40126045
#>  [47,] -0.40229092
#>  [48,]  0.27224959
#>  [49,] -0.02800364
#>  [50,]  0.20261026
#>  [51,]  0.63705980
#>  [52,]  0.01417567
#>  [53,]  0.82358842
#>  [54,]  0.57523621
#>  [55,]  0.60459426
#>  [56,]  1.04843434
#>  [57,] -0.09531194
#>  [58,] -0.74298923
#>  [59,]  0.39382136
#>  [60,]  1.24668154
#>  [61,] -0.24709523
#>  [62,]  0.32260685
#>  [63,]  1.20466563
#>  [64,] -0.85370147
#>  [65,]  0.32003905
#>  [66,] -0.43524868
#>  [67,]  0.91574739
#>  [68,]  1.11361396
#>  [69,]  1.41153464
#>  [70,]  0.96536827
#>  [71,] -1.65425321
#>  [72,] -0.64358452
#>  [73,] -0.82094869
#>  [74,]  0.23362321
#>  [75,] -2.12228630
#>  [76,] -0.38059519
#>  [77,]  0.75461199
#>  [78,] -0.13316243
#>  [79,] -1.09983926
#>  [80,] -0.70002492
#>  [81,]  0.63339198
#>  [82,]  0.08249952
#>  [83,] -0.36895487
#>  [84,] -0.87685639
#>  [85,]  0.20976398
#>  [86,]  0.58212684
#>  [87,]  1.24832474
#>  [88,] -0.02244640
#>  [89,] -0.06193189
#>  [90,]  2.25380389
#>  [91,]  0.09028370
#>  [92,]  1.83798185
#>  [93,] -1.85404206
#>  [94,] -1.65179556
#>  [95,]  0.32064357
#>  [96,] -1.17429970
#>  [97,]  0.81779173
#>  [98,]  0.17926274
#>  [99,]  0.62628784
#> [100,]  0.70518873
#> 
#> $preds
#>                [,1]
#>   [1,] -0.126018865
#>   [2,]  0.352039855
#>   [3,] -0.026917279
#>   [4,] -0.340293533
#>   [5,] -0.312552744
#>   [6,]  0.073987010
#>   [7,] -0.264562503
#>   [8,] -0.041025924
#>   [9,]  0.043070796
#>  [10,] -0.462890299
#>  [11,] -0.229329450
#>  [12,] -0.066972545
#>  [13,] -0.048781994
#>  [14,]  0.002170064
#>  [15,]  0.214835776
#>  [16,]  0.498227459
#>  [17,] -0.331131390
#>  [18,]  0.180361803
#>  [19,]  0.008090857
#>  [20,] -0.200063959
#>  [21,]  0.018511203
#>  [22,]  0.041181697
#>  [23,]  0.265321213
#>  [24,]  0.028245655
#>  [25,] -0.193723154
#>  [26,] -0.036842316
#>  [27,] -0.239164001
#>  [28,]  0.245026914
#>  [29,] -0.177620899
#>  [30,] -0.251118078
#>  [31,]  0.206421695
#>  [32,] -0.008034785
#>  [33,]  0.278937365
#>  [34,] -0.225749543
#>  [35,] -0.342298570
#>  [36,]  0.207653578
#>  [37,]  0.070687253
#>  [38,] -0.142671574
#>  [39,] -0.277704091
#>  [40,]  0.225512834
#>  [41,] -0.388682182
#>  [42,] -0.229301245
#>  [43,]  0.149185057
#>  [44,]  0.039769994
#>  [45,]  0.075440091
#>  [46,] -0.190587740
#>  [47,]  0.129708419
#>  [48,] -0.048315751
#>  [49,] -0.216146296
#>  [50,] -0.047937297
#>  [51,] -0.345241297
#>  [52,] -0.364616744
#>  [53,] -0.195522915
#>  [54,]  0.051575460
#>  [55,] -0.177308497
#>  [56,] -0.098009456
#>  [57,]  0.120286797
#>  [58,]  0.127623460
#>  [59,] -0.050427533
#>  [60,] -0.024187039
#>  [61,]  0.014106286
#>  [62,] -0.041091568
#>  [63,] -0.319346315
#>  [64,] -0.040424612
#>  [65,]  0.123020432
#>  [66,]  0.259242153
#>  [67,] -0.020471817
#>  [68,] -0.359828518
#>  [69,] -0.092596333
#>  [70,]  0.200260975
#>  [71,] -0.398050110
#>  [72,] -0.024312031
#>  [73,] -0.200707394
#>  [74,] -0.014398935
#>  [75,] -0.307767373
#>  [76,] -0.007789670
#>  [77,]  0.054025283
#>  [78,]  0.013099081
#>  [79,] -0.300777277
#>  [80,]  0.155362272
#>  [81,] -0.019450827
#>  [82,]  0.139413647
#>  [83,] -0.437832447
#>  [84,]  0.029718750
#>  [85,] -0.246815242
#>  [86,]  0.007165636
#>  [87,] -0.233171287
#>  [88,]  0.071634415
#>  [89,] -0.187706871
#>  [90,]  0.345336605
#>  [91,]  0.060839753
#>  [92,] -0.180341772
#>  [93,] -0.168434837
#>  [94,]  0.096966192
#>  [95,] -0.077074614
#>  [96,]  0.192371742
#>  [97,]  0.244505665
#>  [98,] -0.231312465
#>  [99,] -0.067566411
#> [100,] -0.135188726
#>