Linear Regression

Estimates a vector for the model coefficients. It also computes model evaluation metrics.

Usage,

our_lm(y, X, alpha = 0.05)

Arguments

y

A vector of length n that represents the response variable of interest that we would like to explain/predict.

X

A matrix of dimension n x p that represents the independent (explanatory) variables.

alpha

A numeric (double) used to denote the significance level.

Value

A list containing the following attributes:

beta

Estimated coefficient vector

sigma2

Estimated variance

variance_beta

Estimated variance for each beta

r_squared

Coefficient of determination

mallow

Mallow's C_p

f_stat

F-statistic

p_value

Determined by F-statistic

ci

Interval of values based on alpha

res

Model Residuals

preds

Model Predictions

Author

Clayton Haley

Ukamaka Victoria Nnyaba

Robert Fett

Examples

x <- matrix(rnorm(400), ncol = 4) # so this is a 100x4 predictor matrix
y <- rnorm(100)
model <- our_lm(y, x, alpha = 0.05)
#> Warning: NaNs produced
#> Warning: NaNs produced
model
#> $beta
#> [1]  0.12032786 -0.22251122 -0.12514800 -0.04057279
#> 
#> $sigma2
#>         [,1]
#> [1,] 3.94862
#> 
#> $variance_beta
#>               [,1]         [,2]         [,3]          [,4]
#> [1,]  0.0415011721 -0.005160579 -0.005554504 -0.0001456576
#> [2,] -0.0051605785  0.045231472  0.002935107  0.0043290334
#> [3,] -0.0055545038  0.002935107  0.030759485  0.0065099337
#> [4,] -0.0001456576  0.004329033  0.006509934  0.0355882943
#> 
#> $r_squared
#> [1] 0.06359865
#> 
#> $mallow
#>          [,1]
#> [1,] 126.3558
#> 
#> $f_stat
#> [1] 2.255729
#> 
#> $p_value
#> [1] 0.08682616
#> 
#> $ci
#>  [1] -0.27895280         NaN         NaN         NaN         NaN -0.63935032
#>  [7] -0.23133224 -0.16952939         NaN -0.32869546 -0.46889398 -0.19871084
#> [13]         NaN -0.35146782 -0.28328605 -0.41031726  0.51960852         NaN
#> [19]         NaN         NaN         NaN  0.19432788 -0.01896377  0.08838380
#> [25]         NaN -0.11632698  0.21859797  0.11756526         NaN -0.09355463
#> [31]  0.03299004  0.32917167
#> 
#> $res
#>               [,1]
#>   [1,]  0.21065600
#>   [2,] -0.34491161
#>   [3,]  0.78287334
#>   [4,] -0.11871181
#>   [5,]  0.20431514
#>   [6,] -0.65490890
#>   [7,]  0.57032942
#>   [8,]  1.08501973
#>   [9,] -0.47639964
#>  [10,]  0.81393656
#>  [11,] -1.69201040
#>  [12,] -0.33531392
#>  [13,] -1.04650819
#>  [14,] -1.38489065
#>  [15,]  0.59747282
#>  [16,]  0.83595692
#>  [17,]  0.56821992
#>  [18,]  0.59497039
#>  [19,]  1.47004419
#>  [20,]  0.29479440
#>  [21,] -1.58774888
#>  [22,]  0.20134738
#>  [23,] -1.98846792
#>  [24,] -0.32926496
#>  [25,] -1.36951710
#>  [26,]  0.51718386
#>  [27,] -0.77851658
#>  [28,]  0.27543127
#>  [29,] -0.14663809
#>  [30,]  2.42455083
#>  [31,] -0.83470624
#>  [32,] -0.28446927
#>  [33,] -1.36982021
#>  [34,] -0.01263741
#>  [35,] -0.87823971
#>  [36,]  0.12258821
#>  [37,] -0.49000291
#>  [38,]  0.63923025
#>  [39,] -2.40315990
#>  [40,] -0.06674165
#>  [41,]  0.85716681
#>  [42,] -2.02644877
#>  [43,] -0.60537606
#>  [44,]  1.37671332
#>  [45,] -0.14520777
#>  [46,]  0.15732224
#>  [47,]  0.64376543
#>  [48,] -0.61895097
#>  [49,] -0.86478518
#>  [50,] -0.43533553
#>  [51,] -0.66675500
#>  [52,]  0.26197188
#>  [53,]  1.09632164
#>  [54,]  1.68882098
#>  [55,] -1.22452701
#>  [56,]  0.68201554
#>  [57,] -0.69925363
#>  [58,] -0.10980463
#>  [59,]  0.62797973
#>  [60,]  0.93830897
#>  [61,] -1.21544854
#>  [62,]  0.53585149
#>  [63,] -0.11727264
#>  [64,] -0.66239693
#>  [65,]  1.63085327
#>  [66,] -0.98422722
#>  [67,] -1.58388515
#>  [68,] -0.67085319
#>  [69,]  1.21422743
#>  [70,] -0.43214530
#>  [71,] -1.07627100
#>  [72,] -0.85911692
#>  [73,] -1.02008454
#>  [74,]  0.03440387
#>  [75,]  0.71689791
#>  [76,] -0.13458904
#>  [77,]  0.97095625
#>  [78,] -0.75275356
#>  [79,] -1.55603762
#>  [80,]  1.37888769
#>  [81,]  0.43353951
#>  [82,]  2.00848281
#>  [83,]  0.79420208
#>  [84,] -0.03878321
#>  [85,]  1.11106701
#>  [86,] -0.41178722
#>  [87,]  0.40090962
#>  [88,] -1.32442036
#>  [89,]  0.56157300
#>  [90,]  1.71901802
#>  [91,]  0.70922246
#>  [92,] -0.09380020
#>  [93,] -2.25754437
#>  [94,] -0.77484508
#>  [95,]  0.98852130
#>  [96,] -0.21086181
#>  [97,] -0.01341704
#>  [98,] -0.27958110
#>  [99,]  0.30177961
#> [100,]  0.35362982
#> 
#> $preds
#>                [,1]
#>   [1,]  0.263477147
#>   [2,] -0.214219179
#>   [3,] -0.067383455
#>   [4,]  0.535213393
#>   [5,]  0.030012599
#>   [6,] -0.058304228
#>   [7,]  0.071165379
#>   [8,]  0.100486485
#>   [9,]  0.129617344
#>  [10,]  0.053309400
#>  [11,]  0.085117734
#>  [12,] -0.282779273
#>  [13,] -0.295265545
#>  [14,] -0.359117302
#>  [15,]  0.228248072
#>  [16,]  0.059395434
#>  [17,]  0.196056848
#>  [18,]  0.026876219
#>  [19,] -0.071958041
#>  [20,]  0.009328448
#>  [21,]  0.049137447
#>  [22,]  0.237263434
#>  [23,]  0.182361847
#>  [24,] -0.509068630
#>  [25,]  0.133195148
#>  [26,]  0.335583686
#>  [27,] -0.141490125
#>  [28,] -0.142978730
#>  [29,]  0.556466012
#>  [30,]  0.201414266
#>  [31,]  0.072445721
#>  [32,] -0.124603987
#>  [33,]  0.269271631
#>  [34,]  0.030643369
#>  [35,] -0.205518893
#>  [36,] -0.451862550
#>  [37,]  0.101599405
#>  [38,] -0.228262876
#>  [39,] -0.378038336
#>  [40,]  0.119572255
#>  [41,]  0.146813616
#>  [42,] -0.110269192
#>  [43,]  0.186940858
#>  [44,] -0.030227649
#>  [45,]  0.461966271
#>  [46,]  0.112820672
#>  [47,] -0.039129263
#>  [48,] -0.182629222
#>  [49,] -0.123085927
#>  [50,]  0.041165432
#>  [51,] -0.135616879
#>  [52,] -0.245873879
#>  [53,] -0.161142015
#>  [54,] -0.189256416
#>  [55,]  0.574731731
#>  [56,]  0.011753500
#>  [57,] -0.234792569
#>  [58,]  0.108117027
#>  [59,] -0.057135332
#>  [60,]  0.764723601
#>  [61,] -0.105150601
#>  [62,] -0.244548451
#>  [63,]  0.580905828
#>  [64,] -0.232728147
#>  [65,]  0.270220190
#>  [66,]  0.341326193
#>  [67,] -0.006560091
#>  [68,]  0.039536503
#>  [69,] -0.032944145
#>  [70,]  0.238744442
#>  [71,]  0.062942091
#>  [72,] -0.192019565
#>  [73,]  0.315822645
#>  [74,] -0.423544450
#>  [75,] -0.194425573
#>  [76,] -0.008664964
#>  [77,]  0.077029821
#>  [78,]  0.027397490
#>  [79,]  0.395401416
#>  [80,]  0.223927435
#>  [81,] -0.147299236
#>  [82,] -0.195789884
#>  [83,]  0.297711781
#>  [84,]  0.367972152
#>  [85,]  0.016835627
#>  [86,]  0.008774796
#>  [87,] -0.128883854
#>  [88,]  0.199203593
#>  [89,]  0.714974992
#>  [90,] -0.384474346
#>  [91,]  0.124492833
#>  [92,]  0.276213740
#>  [93,]  0.338296317
#>  [94,]  0.047265045
#>  [95,]  0.199170621
#>  [96,] -0.284151758
#>  [97,] -0.046578094
#>  [98,] -0.079805983
#>  [99,]  0.279904167
#> [100,] -0.151790013
#>