Linear Regression

Estimates a vector for the model coefficients. It also computes model evaluation metrics.

Usage,

our_lm(y, X, alpha = 0.05)

Arguments

y

A vector of length n that represents the response variable of interest that we would like to explain/predict.

X

A matrix of dimension n x p that represents the independent (explanatory) variables.

alpha

A numeric (double) used to denote the significance level.

Value

A list containing the following attributes:

beta

Estimated coefficient vector

sigma2

Estimated variance

variance_beta

Estimated variance for each beta

r_squared

Coefficient of determination

mallow

Mallow's C_p

f_stat

F-statistic

p_value

Determined by F-statistic

ci

Interval of values based on alpha

res

Model Residuals

preds

Model Predictions

Author

Clayton Haley

Ukamaka Victoria Nnyaba

Robert Fett

Examples

x <- matrix(rnorm(400), ncol = 4) # so this is a 100x4 predictor matrix
y <- rnorm(100)
model <- our_lm(y, x, alpha = 0.05)
#> Warning: NaNs produced
#> Warning: NaNs produced
model
#> $beta
#> [1] -0.02993742  0.09718017 -0.02113102  0.12079249
#> 
#> $sigma2
#>          [,1]
#> [1,] 3.872176
#> 
#> $variance_beta
#>               [,1]          [,2]          [,3]          [,4]
#> [1,]  4.973041e-02 -0.0057317682 -5.475235e-05 -0.0031605518
#> [2,] -5.731768e-03  0.0337545138 -1.848134e-03 -0.0003161652
#> [3,] -5.475235e-05 -0.0018481337  3.194430e-02  0.0008338359
#> [4,] -3.160552e-03 -0.0003161652  8.338359e-04  0.0383009615
#> 
#> $r_squared
#> [1] 0.01623949
#> 
#> $mallow
#>          [,1]
#> [1,] 123.9096
#> 
#> $f_stat
#> [1] 1.423171
#> 
#> $p_value
#> [1] 0.240774
#> 
#> $ci
#>  [1] -0.46701560         NaN         NaN         NaN         NaN -0.26291228
#>  [7]         NaN         NaN         NaN         NaN -0.37143476  0.06419615
#> [13]         NaN         NaN -0.07772736 -0.26278483  0.40714076         NaN
#> [19]         NaN         NaN         NaN  0.45727262         NaN         NaN
#> [25]         NaN         NaN  0.32917273  0.17738884         NaN         NaN
#> [31]  0.03546533  0.50436982
#> 
#> $res
#>                [,1]
#>   [1,] -0.386581156
#>   [2,]  0.497829684
#>   [3,]  0.202657690
#>   [4,] -0.546508065
#>   [5,] -1.345429727
#>   [6,] -0.045276851
#>   [7,] -1.091500201
#>   [8,]  0.279243788
#>   [9,] -1.423636253
#>  [10,] -1.159341620
#>  [11,]  0.638051853
#>  [12,] -0.584243764
#>  [13,]  1.349692391
#>  [14,] -0.379461005
#>  [15,]  0.627740628
#>  [16,] -0.497662072
#>  [17,]  1.000443109
#>  [18,] -0.141228492
#>  [19,]  0.422802170
#>  [20,]  0.110918777
#>  [21,] -0.671971420
#>  [22,]  0.544926055
#>  [23,] -1.055646804
#>  [24,] -0.785211990
#>  [25,]  0.622667082
#>  [26,]  0.357723383
#>  [27,]  0.057177207
#>  [28,]  0.251548800
#>  [29,] -1.138974602
#>  [30,] -0.637270504
#>  [31,] -0.585120262
#>  [32,]  0.038938375
#>  [33,]  0.056463370
#>  [34,]  0.723175595
#>  [35,] -0.949028264
#>  [36,] -0.501281527
#>  [37,] -2.734610811
#>  [38,] -2.746154868
#>  [39,] -1.922205611
#>  [40,] -0.228696317
#>  [41,] -1.311820142
#>  [42,]  0.781004615
#>  [43,]  0.448864695
#>  [44,] -2.304014862
#>  [45,] -0.342190216
#>  [46,]  0.960702645
#>  [47,]  0.772761643
#>  [48,] -0.510470787
#>  [49,] -1.499113166
#>  [50,] -0.505713957
#>  [51,]  0.504632786
#>  [52,]  0.604921087
#>  [53,]  0.344388504
#>  [54,] -0.436875103
#>  [55,] -0.605122600
#>  [56,]  0.225733869
#>  [57,] -1.078323052
#>  [58,] -0.004355873
#>  [59,] -0.364809277
#>  [60,] -0.213395632
#>  [61,]  0.559151112
#>  [62,]  0.494151560
#>  [63,] -2.000700531
#>  [64,]  0.813155431
#>  [65,]  0.084117028
#>  [66,]  0.233788617
#>  [67,]  0.716116553
#>  [68,]  0.493661527
#>  [69,]  2.206917242
#>  [70,] -0.986191112
#>  [71,]  1.631023343
#>  [72,] -2.285084704
#>  [73,] -0.915760300
#>  [74,]  1.615204541
#>  [75,] -1.315567045
#>  [76,] -0.769417632
#>  [77,]  0.747196686
#>  [78,]  0.730813401
#>  [79,]  0.821659338
#>  [80,] -1.286951023
#>  [81,] -0.895184734
#>  [82,]  0.919407683
#>  [83,]  1.291136302
#>  [84,]  0.300560188
#>  [85,]  1.130752920
#>  [86,] -0.461727210
#>  [87,] -0.662438297
#>  [88,] -0.504293238
#>  [89,]  0.641076368
#>  [90,] -0.138621753
#>  [91,] -0.601346714
#>  [92,]  0.659750637
#>  [93,]  0.368011920
#>  [94,]  0.912567149
#>  [95,]  0.147067283
#>  [96,] -0.337433439
#>  [97,]  1.147681091
#>  [98,]  0.314824596
#>  [99,]  1.112270254
#> [100,] -1.077010003
#> 
#> $preds
#>                [,1]
#>   [1,]  0.195523081
#>   [2,]  0.034653126
#>   [3,]  0.261778180
#>   [4,] -0.143743685
#>   [5,]  0.032028126
#>   [6,]  0.169369648
#>   [7,] -0.030167010
#>   [8,]  0.246384005
#>   [9,]  0.086611832
#>  [10,]  0.157294585
#>  [11,]  0.058369807
#>  [12,] -0.097996554
#>  [13,]  0.053420497
#>  [14,]  0.091207880
#>  [15,] -0.205994848
#>  [16,]  0.250861399
#>  [17,]  0.099933254
#>  [18,] -0.112640634
#>  [19,] -0.024711495
#>  [20,] -0.051991600
#>  [21,] -0.177789408
#>  [22,] -0.011395580
#>  [23,]  0.120648166
#>  [24,] -0.064163148
#>  [25,]  0.057286747
#>  [26,] -0.044125364
#>  [27,] -0.272105038
#>  [28,] -0.021597422
#>  [29,]  0.028825561
#>  [30,] -0.181415846
#>  [31,] -0.007759958
#>  [32,]  0.093368828
#>  [33,]  0.189363955
#>  [34,]  0.233152620
#>  [35,] -0.038737836
#>  [36,] -0.057668215
#>  [37,]  0.028802011
#>  [38,] -0.059346751
#>  [39,]  0.079848306
#>  [40,]  0.033149140
#>  [41,]  0.019681925
#>  [42,] -0.063359649
#>  [43,]  0.201184622
#>  [44,]  0.036862375
#>  [45,]  0.100797609
#>  [46,]  0.116343470
#>  [47,] -0.478423402
#>  [48,] -0.088152309
#>  [49,]  0.102734989
#>  [50,] -0.046455681
#>  [51,] -0.113698834
#>  [52,]  0.198845769
#>  [53,]  0.037643950
#>  [54,] -0.026864630
#>  [55,] -0.308653452
#>  [56,] -0.148143039
#>  [57,]  0.391264910
#>  [58,] -0.390556946
#>  [59,]  0.150535753
#>  [60,]  0.289854563
#>  [61,]  0.297982021
#>  [62,]  0.160598701
#>  [63,] -0.001390429
#>  [64,]  0.153444845
#>  [65,]  0.004905571
#>  [66,]  0.109815557
#>  [67,] -0.131005921
#>  [68,] -0.205495911
#>  [69,] -0.011505630
#>  [70,] -0.052454232
#>  [71,] -0.081488161
#>  [72,] -0.058795927
#>  [73,]  0.294891851
#>  [74,] -0.125893085
#>  [75,]  0.083959138
#>  [76,] -0.105864682
#>  [77,]  0.212093110
#>  [78,]  0.097642790
#>  [79,] -0.059972801
#>  [80,] -0.113028753
#>  [81,]  0.182839700
#>  [82,] -0.022208974
#>  [83,]  0.011856522
#>  [84,]  0.030998680
#>  [85,]  0.224166689
#>  [86,]  0.072677171
#>  [87,]  0.067056098
#>  [88,] -0.387128122
#>  [89,]  0.439502878
#>  [90,]  0.122638837
#>  [91,]  0.011415850
#>  [92,] -0.068064961
#>  [93,]  0.034776853
#>  [94,]  0.023129447
#>  [95,]  0.290375719
#>  [96,] -0.203399721
#>  [97,] -0.115008505
#>  [98,] -0.023333125
#>  [99,]  0.018995506
#> [100,] -0.135512382
#>