Linear Regression

Estimates a vector for the model coefficients. It also computes model evaluation metrics.

Usage,

our_lm(y, X, alpha = 0.05)

Arguments

y

A vector of length n that represents the response variable of interest that we would like to explain/predict.

X

A matrix of dimension n x p that represents the independent (explanatory) variables.

alpha

A numeric (double) used to denote the significance level.

Value

A list containing the following attributes:

beta

Estimated coefficient vector

sigma2

Estimated variance

variance_beta

Estimated variance for each beta

r_squared

Coefficient of determination

mallow

Mallow's C_p

f_stat

F-statistic

p_value

Determined by F-statistic

ci

Interval of values based on alpha

res

Model Residuals

preds

Model Predictions

Author

Clayton Haley

Ukamaka Victoria Nnyaba

Robert Fett

Examples

x <- matrix(rnorm(400), ncol = 4) # so this is a 100x4 predictor matrix
y <- rnorm(100)
model <- our_lm(y, x, alpha = 0.05)
#> Warning: NaNs produced
#> Warning: NaNs produced
model
#> $beta
#> [1]  0.04375043 -0.13296651 -0.02416402  0.08668004
#> 
#> $sigma2
#>          [,1]
#> [1,] 4.417666
#> 
#> $variance_beta
#>               [,1]         [,2]         [,3]          [,4]
#> [1,]  0.0404351615 -0.004008644 -0.002046384 -0.0006104362
#> [2,] -0.0040086440  0.058163232 -0.006126954 -0.0016078972
#> [3,] -0.0020463844 -0.006126954  0.050334288  0.0057518891
#> [4,] -0.0006104362 -0.001607897  0.005751889  0.0437943206
#> 
#> $r_squared
#> [1] 0.01549544
#> 
#> $mallow
#>          [,1]
#> [1,] 141.3653
#> 
#> $f_stat
#> [1] 0.8481585
#> 
#> $p_value
#> [1] 0.4709094
#> 
#> $ci
#>  [1] -0.3503689        NaN        NaN        NaN        NaN -0.6056521
#>  [7]        NaN        NaN        NaN        NaN -0.4638879 -0.0619660
#> [13]        NaN        NaN -0.1728101 -0.3234834  0.4378697        NaN
#> [19]        NaN        NaN        NaN  0.3397191        NaN        NaN
#> [25]        NaN        NaN  0.4155599  0.2353261        NaN        NaN
#> [31]  0.1244820  0.4968435
#> 
#> $res
#>               [,1]
#>   [1,]  0.17872512
#>   [2,] -0.12202231
#>   [3,]  0.28938841
#>   [4,] -0.83074832
#>   [5,]  2.44996576
#>   [6,] -1.93154624
#>   [7,]  0.06552260
#>   [8,] -1.46893667
#>   [9,]  0.73087684
#>  [10,] -1.11821651
#>  [11,]  0.25099029
#>  [12,] -1.35283858
#>  [13,]  1.95234449
#>  [14,] -0.48143581
#>  [15,]  0.90403407
#>  [16,]  0.13814267
#>  [17,]  0.45877117
#>  [18,] -1.23274988
#>  [19,]  0.67322690
#>  [20,]  0.87217058
#>  [21,] -0.96398198
#>  [22,] -0.49161039
#>  [23,] -1.39098689
#>  [24,]  0.97240905
#>  [25,]  0.16625699
#>  [26,] -0.12419089
#>  [27,] -0.40684904
#>  [28,] -0.55241769
#>  [29,]  0.20909193
#>  [30,]  0.04966795
#>  [31,]  1.57392443
#>  [32,]  0.22340723
#>  [33,]  0.32888434
#>  [34,] -1.65571753
#>  [35,]  0.50049369
#>  [36,] -1.06208282
#>  [37,] -0.87416708
#>  [38,]  0.53904099
#>  [39,] -1.46844104
#>  [40,] -1.25492249
#>  [41,]  0.45287147
#>  [42,]  0.49180601
#>  [43,]  0.66160048
#>  [44,] -0.52050795
#>  [45,] -0.79959385
#>  [46,] -0.97363518
#>  [47,]  0.94703982
#>  [48,]  0.24038701
#>  [49,] -0.06356824
#>  [50,] -1.41748685
#>  [51,]  1.66209323
#>  [52,]  1.74461610
#>  [53,] -1.12931284
#>  [54,] -0.09365893
#>  [55,] -0.30820483
#>  [56,] -0.15937152
#>  [57,] -1.72990109
#>  [58,]  1.02968236
#>  [59,]  0.52647044
#>  [60,] -1.24811916
#>  [61,]  0.26534427
#>  [62,] -0.75579809
#>  [63,]  0.80072460
#>  [64,] -1.14598728
#>  [65,] -0.14761666
#>  [66,]  0.63420850
#>  [67,] -0.10889671
#>  [68,] -1.87581706
#>  [69,]  1.63175676
#>  [70,]  0.41229023
#>  [71,] -1.23554639
#>  [72,]  1.93602778
#>  [73,] -0.03885511
#>  [74,]  0.44459447
#>  [75,] -1.16653982
#>  [76,]  0.45609736
#>  [77,]  1.67281037
#>  [78,]  2.09582400
#>  [79,] -0.08708378
#>  [80,] -0.54425230
#>  [81,]  1.33106366
#>  [82,] -2.85028863
#>  [83,] -0.05287229
#>  [84,]  1.49515900
#>  [85,]  0.56116563
#>  [86,]  0.51149651
#>  [87,] -0.06677334
#>  [88,]  0.65084782
#>  [89,] -0.19906880
#>  [90,] -0.71590304
#>  [91,] -0.37012268
#>  [92,] -1.31439582
#>  [93,]  0.32260403
#>  [94,] -0.86306874
#>  [95,] -1.50346331
#>  [96,]  1.18651646
#>  [97,]  0.28329420
#>  [98,] -1.28064364
#>  [99,] -1.05245011
#> [100,] -0.15247619
#> 
#> $preds
#>                [,1]
#>   [1,]  0.058848029
#>   [2,] -0.168314933
#>   [3,] -0.039855321
#>   [4,] -0.253390486
#>   [5,] -0.012782203
#>   [6,]  0.020722163
#>   [7,] -0.039126075
#>   [8,] -0.062457892
#>   [9,]  0.003546352
#>  [10,] -0.189877476
#>  [11,] -0.305579726
#>  [12,] -0.023419818
#>  [13,]  0.271839515
#>  [14,]  0.007397198
#>  [15,] -0.027840495
#>  [16,]  0.181434943
#>  [17,]  0.197666907
#>  [18,]  0.056314883
#>  [19,] -0.138285844
#>  [20,] -0.112261948
#>  [21,] -0.277512516
#>  [22,] -0.140073785
#>  [23,]  0.036476312
#>  [24,]  0.020278898
#>  [25,]  0.104376187
#>  [26,] -0.065599352
#>  [27,] -0.100179956
#>  [28,]  0.140847981
#>  [29,]  0.320222705
#>  [30,] -0.172556341
#>  [31,] -0.072796291
#>  [32,] -0.054059954
#>  [33,] -0.350686403
#>  [34,]  0.126920086
#>  [35,]  0.219364319
#>  [36,] -0.016802414
#>  [37,] -0.053207138
#>  [38,] -0.033971482
#>  [39,]  0.087544569
#>  [40,] -0.091702602
#>  [41,] -0.095812496
#>  [42,] -0.157621783
#>  [43,]  0.007516100
#>  [44,]  0.034100251
#>  [45,]  0.212290957
#>  [46,] -0.303496905
#>  [47,] -0.041432480
#>  [48,] -0.052597569
#>  [49,] -0.018428143
#>  [50,]  0.064215999
#>  [51,] -0.032635622
#>  [52,] -0.066079083
#>  [53,] -0.133709812
#>  [54,]  0.190627660
#>  [55,] -0.029106199
#>  [56,] -0.312108275
#>  [57,] -0.033220088
#>  [58,]  0.032323891
#>  [59,]  0.011417452
#>  [60,]  0.079289984
#>  [61,] -0.624925169
#>  [62,]  0.356615264
#>  [63,]  0.058632135
#>  [64,] -0.028639399
#>  [65,] -0.026142599
#>  [66,] -0.222166763
#>  [67,]  0.185215169
#>  [68,] -0.066270419
#>  [69,] -0.076931644
#>  [70,] -0.072841647
#>  [71,] -0.009792349
#>  [72,] -0.015339586
#>  [73,]  0.055527949
#>  [74,]  0.047512995
#>  [75,]  0.030916600
#>  [76,] -0.145210481
#>  [77,] -0.060742565
#>  [78,]  0.117026736
#>  [79,]  0.187996797
#>  [80,] -0.160097738
#>  [81,]  0.132204939
#>  [82,]  0.290578447
#>  [83,] -0.095657908
#>  [84,]  0.089845782
#>  [85,]  0.023713587
#>  [86,]  0.119406992
#>  [87,] -0.218394806
#>  [88,] -0.020647477
#>  [89,]  0.118655362
#>  [90,] -0.296037478
#>  [91,]  0.107398789
#>  [92,] -0.033809182
#>  [93,]  0.160335024
#>  [94,]  0.004770839
#>  [95,] -0.017057994
#>  [96,]  0.001135071
#>  [97,] -0.030443908
#>  [98,] -0.128060391
#>  [99,]  0.233777133
#> [100,] -0.021054715
#>