Linear Regression

Estimates a vector for the model coefficients. It also computes model evaluation metrics.

Usage,

our_lm(y, X, alpha = 0.05)

Arguments

y

A vector of length n that represents the response variable of interest that we would like to explain/predict.

X

A matrix of dimension n x p that represents the independent (explanatory) variables.

alpha

A numeric (double) used to denote the significance level.

Value

A list containing the following attributes:

beta

Estimated coefficient vector

sigma2

Estimated variance

variance_beta

Estimated variance for each beta

r_squared

Coefficient of determination

mallow

Mallow's C_p

f_stat

F-statistic

p_value

Determined by F-statistic

ci

Interval of values based on alpha

res

Model Residuals

preds

Model Predictions

Author

Clayton Haley

Ukamaka Victoria Nnyaba

Robert Fett

Examples

x <- matrix(rnorm(400), ncol = 4) # so this is a 100x4 predictor matrix
y <- rnorm(100)
model <- our_lm(y, x, alpha = 0.05)
#> Warning: NaNs produced
#> Warning: NaNs produced
model
#> $beta
#> [1]  0.07220380 -0.04477389 -0.02831995  0.04606931
#> 
#> $sigma2
#>          [,1]
#> [1,] 4.030012
#> 
#> $variance_beta
#>               [,1]         [,2]          [,3]         [,4]
#> [1,]  0.0413812971 -0.003315245  0.0003556898 0.0043245271
#> [2,] -0.0033152454  0.048525599 -0.0028689024 0.0043940616
#> [3,]  0.0003556898 -0.002868902  0.0506749946 0.0005074664
#> [4,]  0.0043245271  0.004394062  0.0005074664 0.0479453584
#> 
#> $r_squared
#> [1] 0.006684122
#> 
#> $mallow
#>          [,1]
#> [1,] 128.9604
#> 
#> $f_stat
#> [1] 0.3156977
#> 
#> $p_value
#> [1] 0.8139923
#> 
#> $ci
#>  [1] -0.326499787          NaN -0.065284361 -0.082820147          NaN
#>  [6] -0.476525083          NaN -0.083852228  0.035239389          NaN
#> [11] -0.469529544  0.001917174 -0.056685661 -0.174695431 -0.072472089
#> [16] -0.383092802  0.470907383          NaN  0.008644458  0.174958771
#> [21]          NaN  0.386977301          NaN  0.175990852  0.109168207
#> [26]          NaN  0.412889642  0.090221450  0.201093257  0.085147649
#> [31]  0.015832187  0.475231426
#> 
#> $res
#>                [,1]
#>   [1,] -1.455393772
#>   [2,] -0.406750483
#>   [3,]  0.145080824
#>   [4,] -0.081441393
#>   [5,] -0.336009971
#>   [6,] -0.905063112
#>   [7,] -0.026917763
#>   [8,]  0.925223576
#>   [9,]  0.516497949
#>  [10,] -1.843700999
#>  [11,] -0.035469782
#>  [12,]  1.308957204
#>  [13,] -0.990703574
#>  [14,]  1.175738900
#>  [15,] -0.696864418
#>  [16,]  0.160614656
#>  [17,]  0.809923274
#>  [18,]  1.789993753
#>  [19,]  0.141557045
#>  [20,]  0.367287100
#>  [21,] -0.935778898
#>  [22,]  1.046730892
#>  [23,]  1.258961389
#>  [24,]  1.315793154
#>  [25,]  1.895211658
#>  [26,] -1.455831374
#>  [27,]  0.594848733
#>  [28,] -0.424912902
#>  [29,]  0.700420884
#>  [30,] -0.078853124
#>  [31,] -0.232954200
#>  [32,]  0.752649452
#>  [33,] -0.440514132
#>  [34,] -0.195907259
#>  [35,]  1.010857197
#>  [36,] -2.236238409
#>  [37,]  0.156210041
#>  [38,]  0.187342119
#>  [39,]  1.131204991
#>  [40,]  0.140685825
#>  [41,] -0.914328551
#>  [42,]  1.399316655
#>  [43,] -0.115975202
#>  [44,]  1.400145057
#>  [45,]  0.797958484
#>  [46,]  0.686936145
#>  [47,]  0.665918751
#>  [48,] -1.562835830
#>  [49,]  0.022930034
#>  [50,] -1.137798402
#>  [51,] -0.693300061
#>  [52,]  0.303506457
#>  [53,] -0.305639968
#>  [54,] -0.003486789
#>  [55,] -0.804387438
#>  [56,] -0.117698317
#>  [57,]  1.880756010
#>  [58,]  0.860111644
#>  [59,] -0.253328329
#>  [60,] -1.428579203
#>  [61,] -2.868373570
#>  [62,]  1.702296631
#>  [63,]  0.700778542
#>  [64,]  0.685950426
#>  [65,] -0.006019938
#>  [66,] -0.997423518
#>  [67,]  1.949256476
#>  [68,]  0.500718364
#>  [69,] -0.863831422
#>  [70,]  0.735367180
#>  [71,]  2.379542539
#>  [72,] -0.494263761
#>  [73,] -0.710345245
#>  [74,] -2.127731448
#>  [75,]  0.939215647
#>  [76,]  0.322884420
#>  [77,] -0.384110012
#>  [78,] -0.810162728
#>  [79,]  0.149290268
#>  [80,]  0.290025206
#>  [81,] -0.535296090
#>  [82,]  1.222380754
#>  [83,] -0.459191505
#>  [84,] -0.681507376
#>  [85,]  0.652609466
#>  [86,] -1.132863041
#>  [87,]  0.146510790
#>  [88,] -1.300987319
#>  [89,] -0.334508086
#>  [90,] -1.118449516
#>  [91,] -0.336538403
#>  [92,]  0.140608166
#>  [93,] -1.319618545
#>  [94,]  1.101203850
#>  [95,] -0.325538658
#>  [96,]  0.538049986
#>  [97,]  0.111475929
#>  [98,] -0.316721194
#>  [99,]  0.294353857
#> [100,]  0.373414763
#> 
#> $preds
#>                [,1]
#>   [1,]  0.085629540
#>   [2,] -0.118929111
#>   [3,] -0.019800341
#>   [4,]  0.058343790
#>   [5,] -0.081955897
#>   [6,]  0.031897461
#>   [7,]  0.123714583
#>   [8,]  0.022072386
#>   [9,] -0.027949810
#>  [10,] -0.073070603
#>  [11,] -0.177817229
#>  [12,]  0.042454021
#>  [13,] -0.037276302
#>  [14,]  0.056737220
#>  [15,]  0.174962707
#>  [16,] -0.005064944
#>  [17,]  0.017005934
#>  [18,]  0.011957080
#>  [19,] -0.032082309
#>  [20,] -0.053256195
#>  [21,]  0.169331841
#>  [22,] -0.065863412
#>  [23,]  0.024537960
#>  [24,] -0.095795060
#>  [25,]  0.013514497
#>  [26,]  0.042881853
#>  [27,] -0.047400457
#>  [28,] -0.063104081
#>  [29,] -0.021668383
#>  [30,]  0.076525562
#>  [31,]  0.009515743
#>  [32,]  0.025461785
#>  [33,]  0.100924934
#>  [34,]  0.005504974
#>  [35,] -0.059775023
#>  [36,] -0.070133289
#>  [37,] -0.055845176
#>  [38,] -0.081940050
#>  [39,]  0.005729216
#>  [40,] -0.021100401
#>  [41,]  0.010186616
#>  [42,] -0.097009296
#>  [43,]  0.015879144
#>  [44,]  0.128120656
#>  [45,] -0.043413494
#>  [46,]  0.198961541
#>  [47,]  0.100664572
#>  [48,] -0.171723498
#>  [49,]  0.117798856
#>  [50,]  0.127835461
#>  [51,]  0.044182288
#>  [52,]  0.095096413
#>  [53,]  0.070335357
#>  [54,]  0.007502234
#>  [55,]  0.128527704
#>  [56,]  0.145533811
#>  [57,]  0.090885076
#>  [58,] -0.066002473
#>  [59,] -0.030983194
#>  [60,] -0.045948879
#>  [61,]  0.070694865
#>  [62,]  0.103655996
#>  [63,] -0.039633212
#>  [64,] -0.028452922
#>  [65,] -0.055864359
#>  [66,] -0.002102273
#>  [67,] -0.070361625
#>  [68,]  0.144552399
#>  [69,] -0.045493060
#>  [70,]  0.108836159
#>  [71,]  0.102552386
#>  [72,]  0.169267645
#>  [73,] -0.184006775
#>  [74,]  0.068903280
#>  [75,] -0.139727923
#>  [76,] -0.006076047
#>  [77,] -0.011470134
#>  [78,]  0.046321728
#>  [79,]  0.105738790
#>  [80,]  0.055809004
#>  [81,]  0.040344660
#>  [82,]  0.046776562
#>  [83,] -0.089392848
#>  [84,]  0.200436132
#>  [85,]  0.048846721
#>  [86,]  0.041252350
#>  [87,]  0.136724636
#>  [88,] -0.011788316
#>  [89,] -0.113125035
#>  [90,]  0.232985407
#>  [91,] -0.181540697
#>  [92,] -0.023213212
#>  [93,] -0.057227106
#>  [94,]  0.035469835
#>  [95,]  0.219352204
#>  [96,] -0.048478691
#>  [97,]  0.004278761
#>  [98,] -0.165911914
#>  [99,]  0.094060641
#> [100,] -0.077328733
#>