Linear Regression

Estimates a vector for the model coefficients. It also computes model evaluation metrics.

Usage,

our_lm(y, x, alpha = 0.05)

Arguments

y

A vector of length n that represents the response variable of interest that we would like to explain/predict.

alpha

A numeric (double) used to denote the significance level.

X

A matrix of dimension n x p that represents the independent (explanatory) variables.

Value

A list containing the following attributes:

beta

Estimated coefficient vector

sigma2

Estimated variance

variance_beta

Estimated variance for each beta

r_squared

Coefficient of determination

mallow

Mallow's C_p

f_stat

F-statistic

p_value

Determined by F-statistic

ci

Confidence Intervals based on alpha

res

Model Residuals

preds

Model Predictions

Author

Clayton Haley

Ukamaka Victoria Nnyaba

Robert Fett

Examples

x <- matrix(rnorm(400), ncol = 4) # so this is a 100x4 predictor matrix
y <- rnorm(100)
model <- our_lm(y, x, alpha = 0.05)
#> Warning: NaNs produced
#> Warning: NaNs produced
model
#> $beta
#> [1] -0.176446616  0.147412983 -0.245922892 -0.009864657 -0.029704275
#> 
#> $sigma2
#>          [,1]
#> [1,] 6.338869
#> 
#> $variance_beta
#>               [,1]          [,2]          [,3]          [,4]         [,5]
#> [1,]  6.435803e-02  1.490965e-05 -2.730766e-03 -6.586616e-03 -0.001030474
#> [2,]  1.490965e-05  7.063961e-02 -1.405795e-02 -3.335092e-03 -0.005212279
#> [3,] -2.730766e-03 -1.405795e-02  6.795718e-02 -2.383246e-05  0.005046672
#> [4,] -6.586616e-03 -3.335092e-03 -2.383246e-05  5.186558e-02  0.001129990
#> [5,] -1.030474e-03 -5.212279e-03  5.046672e-03  1.129990e-03  0.063033444
#> 
#> $r_squared
#> [1] 0.05134304
#> 
#> $mallow
#>          [,1]
#> [1,] 183.8272
#> 
#> $f_stat
#> [1] 1.28567
#> 
#> $p_value
#> [1] 0.2811435
#> 
#> $ci
#>  [1] -0.67366762  0.13984497         NaN         NaN         NaN -0.18401463
#>  [7] -0.37350845         NaN         NaN         NaN         NaN         NaN
#> [13] -0.75685801         NaN -0.16893999         NaN         NaN         NaN
#> [19] -0.45622715 -0.09558910         NaN         NaN -0.38515860 -0.07574948
#> [25] -0.52178189  0.32077439  0.15498100         NaN         NaN         NaN
#> [31] -0.16887860  0.66833441         NaN         NaN         NaN         NaN
#> [37]         NaN  0.26501223         NaN  0.10953144         NaN         NaN
#> [43]         NaN  0.43649783  0.03618055         NaN         NaN -0.10668718
#> [49]  0.05602016  0.46237334
#> 
#> $res
#>               [,1]
#>   [1,]  0.20628750
#>   [2,]  0.01014647
#>   [3,] -0.39127916
#>   [4,]  0.30163649
#>   [5,] -0.52871806
#>   [6,]  0.49754165
#>   [7,] -0.18865966
#>   [8,] -0.28195759
#>   [9,] -0.27500500
#>  [10,]  0.34140318
#>  [11,] -0.66799369
#>  [12,]  0.49833044
#>  [13,]  2.25879296
#>  [14,] -0.47749257
#>  [15,] -0.05402087
#>  [16,] -1.17590626
#>  [17,]  0.45761597
#>  [18,]  1.62441074
#>  [19,] -0.08436815
#>  [20,]  0.60861364
#>  [21,]  0.96138672
#>  [22,]  1.70345475
#>  [23,] -0.44381868
#>  [24,]  0.96651494
#>  [25,]  1.09691576
#>  [26,] -0.39229748
#>  [27,] -0.46174465
#>  [28,]  0.95371542
#>  [29,] -2.05197747
#>  [30,] -0.55740190
#>  [31,]  1.14886106
#>  [32,]  0.03332998
#>  [33,]  0.58398315
#>  [34,] -0.57194528
#>  [35,]  2.01749479
#>  [36,]  0.19330625
#>  [37,] -0.43626868
#>  [38,] -1.09165175
#>  [39,]  0.28813871
#>  [40,]  0.55002879
#>  [41,] -0.98775519
#>  [42,] -0.61662591
#>  [43,]  0.66255880
#>  [44,] -0.82395296
#>  [45,]  0.43388263
#>  [46,]  0.66529083
#>  [47,] -2.02148016
#>  [48,] -0.29093435
#>  [49,]  0.90839459
#>  [50,] -1.48291204
#>  [51,] -1.05672468
#>  [52,] -0.13944952
#>  [53,] -0.15261449
#>  [54,] -1.44578578
#>  [55,]  0.07861999
#>  [56,]  0.68731096
#>  [57,]  0.37339286
#>  [58,]  0.08763049
#>  [59,] -2.37868331
#>  [60,]  0.52900375
#>  [61,]  1.60095901
#>  [62,] -1.08113190
#>  [63,]  0.25941772
#>  [64,] -0.68416659
#>  [65,] -0.12722814
#>  [66,]  0.48046496
#>  [67,] -0.93864117
#>  [68,]  2.28816792
#>  [69,]  0.81471885
#>  [70,] -0.48019498
#>  [71,] -0.81276372
#>  [72,] -1.43571132
#>  [73,] -1.73439062
#>  [74,] -3.14995681
#>  [75,] -1.09557105
#>  [76,] -0.38263234
#>  [77,]  0.65507350
#>  [78,]  3.75465122
#>  [79,] -2.30472307
#>  [80,] -0.54468452
#>  [81,] -0.92177813
#>  [82,]  0.28803761
#>  [83,]  1.35986526
#>  [84,] -0.53417010
#>  [85,]  1.39791742
#>  [86,]  0.04886367
#>  [87,] -1.13852427
#>  [88,] -0.70079661
#>  [89,]  0.78537544
#>  [90,]  0.34547691
#>  [91,] -0.51179103
#>  [92,]  2.60685323
#>  [93,]  0.93308536
#>  [94,]  1.53688356
#>  [95,] -0.62152204
#>  [96,]  0.54288030
#>  [97,] -0.13595582
#>  [98,]  0.70333572
#>  [99,] -0.98978443
#> [100,] -0.27333072
#> 
#> $preds
#>                [,1]
#>   [1,] -0.210205742
#>   [2,]  0.211268332
#>   [3,] -0.386281464
#>   [4,] -0.492030298
#>   [5,]  0.271879934
#>   [6,] -0.025984831
#>   [7,] -0.138400678
#>   [8,] -0.066113068
#>   [9,]  0.395608585
#>  [10,] -0.254889084
#>  [11,] -0.324179004
#>  [12,] -0.240918426
#>  [13,] -0.007510675
#>  [14,] -0.091655178
#>  [15,] -0.544886295
#>  [16,] -0.143844069
#>  [17,] -0.649455294
#>  [18,]  0.052916360
#>  [19,] -0.006581338
#>  [20,] -0.479752122
#>  [21,]  0.144239586
#>  [22,] -0.619006199
#>  [23,] -0.383978672
#>  [24,] -0.371781568
#>  [25,] -0.495236528
#>  [26,]  0.565674983
#>  [27,] -0.330051383
#>  [28,] -0.175855995
#>  [29,] -0.192353915
#>  [30,] -0.588080279
#>  [31,] -0.176285794
#>  [32,] -0.183573975
#>  [33,]  0.042772688
#>  [34,] -0.194926466
#>  [35,] -0.281695911
#>  [36,] -0.181535028
#>  [37,] -0.151027213
#>  [38,] -0.491809819
#>  [39,] -0.572608376
#>  [40,] -0.637788727
#>  [41,] -0.299021923
#>  [42,] -0.061200262
#>  [43,] -0.309346327
#>  [44,] -0.143677032
#>  [45,]  0.172306165
#>  [46,] -0.115905839
#>  [47,] -0.140600744
#>  [48,] -0.187120670
#>  [49,] -0.550568048
#>  [50,]  0.453735157
#>  [51,]  0.144206220
#>  [52,]  0.110517846
#>  [53,] -0.286723891
#>  [54,] -0.616603114
#>  [55,]  0.112097208
#>  [56,] -0.204880981
#>  [57,] -0.281335922
#>  [58,] -0.311352569
#>  [59,] -0.424992989
#>  [60,]  0.109021623
#>  [61,]  0.217890149
#>  [62,]  0.256006270
#>  [63,]  0.075030006
#>  [64,] -0.142852453
#>  [65,] -0.419805545
#>  [66,] -0.416541323
#>  [67,] -0.396153572
#>  [68,] -0.311067255
#>  [69,] -0.479717047
#>  [70,] -0.062223855
#>  [71,]  0.096084299
#>  [72,] -0.161925169
#>  [73,] -0.083619924
#>  [74,] -0.849103798
#>  [75,] -0.013072424
#>  [76,] -0.509170126
#>  [77,] -0.479174820
#>  [78,] -0.386973348
#>  [79,] -0.393304551
#>  [80,] -0.244731172
#>  [81,] -0.191675288
#>  [82,] -0.391539635
#>  [83,] -0.129601465
#>  [84,] -0.193649706
#>  [85,]  0.189168435
#>  [86,]  0.028336054
#>  [87,] -0.274481150
#>  [88,] -0.268253926
#>  [89,] -0.241869491
#>  [90,]  0.038349632
#>  [91,] -0.084501536
#>  [92,] -0.255994162
#>  [93,] -0.078089310
#>  [94,] -0.248625341
#>  [95,] -0.015941281
#>  [96,] -0.228068692
#>  [97,] -0.173935397
#>  [98,] -0.008774484
#>  [99,] -0.069926299
#> [100,] -0.086454332
#>